Set minimal memory limit for CirrOs image
TestVM (that is cirros) is installed by fuel
in each new openstack cloud. This image doesn't
have any limits set in openstack. If user tries
to launch this image on flavor with 16Mb of
memory it will go into infinite boot. It would be
good to prevent such accidental hard-to-find
issues by setting limits for image on required
resources.

Change-Id: I37c917c43f89dd37dd07e35f9a7d0379e5f3a72f
Related-Bug: #1304664