Unlock nailgun-agent on failed nodes (discovery process)
If for some reason reboot nodes from boostrap to provisioning
state fail, nodes will disabled for discovery. This action has
affect: Nailgun set them offline status and prevent any
operation with them.

Now if nodes online, we will show this and give a chance to
retry operation.

This is not the direct fix, but decrise potencial
damage.

Change-Id: I6df881be61c038a1c108526c6defe00d2c5d0452
Related-Bug: #1336247