Set no quorum policy stop at post deployment
Corosync cluster should being deployed with 'ignore' policy
and puppet ensures that. Later, if there are 3 or more controllers
in cluster, its policy should be updated to 'stop'
(or 'suicide', if fencing enabled) in order to handle split brain
scenarios as appropriate. Otherwise, in case of there are 2 or less
controllers in cluster, qourum policy should not be changed.

Closes-bug: #1348548

Change-Id: Iae2df7736992edcef4ae72f967af9f504d56612b
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>