Repeated erase and reboot for bootstrap nodes
If Atute got unxepecting status 'bootstrap'
from node in the middle of provision operation,
we try to erase and reboot node. In mostly
enviroments where this bug was reproduced
this operation was enough to solve problem.

If, for some reason, this does not help, we
can with great confidence to look at the
problem in Cobbler.

Change-Id: I9ffaa885bd150a6fa119f07bc04b79bd5afb92a1
Closes-Bug: #1318567