Stop corrupting the provisioned nodes' filesystems.
As of now astute abruptly reboots the nodes which have been just provisioned.
(basically it sends SysRq-b via ssh). Fortunately this command does not reach
most of the nodes: before shooting a node with SysRq-b astute tries to reboot
it gracefully via cobbler. Thus most of the time the ssh connection which
tries to send SysRq-b gets rejected. However if a node reboots fast enough
the `control reboot' hits such a node (which is being configured by cloud-init
at that time). An abrupt reboot causes a major filesystem corruption and
makes the deployment fail. All in all the control reboot thing causes much
worse problem than the one it tries to resolve. Disable the control reboot
for now.

Related-bug: #1394599
Related-bug: #1407634
Change-Id: Ie4af6904a8297d9acbc4e96425903e9e57450286