Fix rebooting of the bootstrap nodes
Skip the hard reboot for the image based provisioning since the reboot
command might hit a node which has booted into the provisioned OS (which
causes the filesystem corruption and interrupts the deployment).
Fix the condition which selects the bootstrap nodes, that is, use
SshHardReboot instead of SshRebootNotProvisioning (the latter reboots
the locally booted nodes instead the bootstrap ones due to the inverted
condition).

Related-bug: #1394599
Related-bug: #1407634
Change-Id: Ie4af6904a8297d9acbc4e96425903e9e57450286