IBP: Added sleep 10 after node reboot
The reason for this fix to appear is that Cobbler runs reboot
command and reports reboot succeeded, but astute manages to get the type
of node OS via Mcollective (/etc/nailgun_systemtype) and runs node erasing
procedure for the nodes which type!=target. Erasing procedure in turn
kicks sysrq trigger which breaks the node.

Change-Id: I98cc29256fb9a807aba9b6dddc500864c74bd735
Closes-Bug: 1420256