Added handling of mclient_remove parameter
If mclient_remove parameter is set to false for a node, it causes mcollective
client not to be called for this node. This is because we want to be able to
forcefuly delete offline nodes and without this parameter Astute keeps
retrying deletion for a long time.
If mclient_remove is not specified for node, it is treated as true by default
thus preserving backwards compatibility.

Implements blueprint: node-removal-task

Change-Id: I256ebdb731a405897192f37b874c6d4f785e3e45