Don't use execute_tasks for pre_provision tasks
The `execute_tasks` method was designed to execute some abstract tasks
sent by Nailgun. Hence, in the end of execution if all is ok it sends
the "ready" status back to Nailgun. It's wrong in case of pre_provision
tasks, because they don't mean provisioning is done.

Since this commit we're running pre_provision tasks by means Nailgun
Hooks and therefore do not report "ready" when they are done.

Closes-Bug: #1433060
Change-Id: I9d0cf11b1788c85937b5a15e44fc904cb24efbe3