Erase non-root disks first
This change adjusts the order that the system block devices (disks)
are erased as part erase_node MCollective action.  This change queries
the block devices for the system and uses lsblk to determine if the
device has a partition that is currently being used for mounting /.
The disks that are not used for / are erased first, followed by
the root disks.  This done to ensure that data disks are properly
cleaned up and are not accidentally left intact on the chance that
purging the root disk causes the erase script to fail.

Change-Id: I17e0e88224ef781b666b100f80904e5abfecd023
Closes-Bug: 1437511