Prevented failed nodes from being rebooted
Provision fault tolerance makes it possible
to continue deployment even if some of nodes
failed while being provisioned. But if
reboot such failed node with chain loader
it likely can stick on this stage and we won't
be able to reboot it any way except turning it off
and on.

This change prevents failed nodes from being rebooted
with chain loader. Besides this patch sets
profile for IBP nodes into 'bootstrap', so as to
avoid them being booted with anaconda or debian-installer.

Change-Id: Id2b500d359fa2772ef0a81bef37f2f7152894b00
Closes-Bug: #1455385