Removed sync from reboot_nodes
The thing is we reboot nodes one by one
sleeping couple seconds between subsequent reboot RPC calls.
Sometimes we spend quite long time doing this, time
which is enough for one or more nodes to reboot.
We then call sync method (which is absolutely unnecessary)
and cobbler removes all tftp configs and creates them
again. And sometimes a node can not get the config that
makes it stick on boot asking a user to press any key.

Change-Id: I3ac5a887725411abdfbe1209b8e9f986fbb62666
Closes-Bug: #1456247