Moved nodiscover to /etc/nailgun-agent
We create /etc/nailgun-agent/nodiscover during
provisioning and then this file is assumed to be
removed in /etc/rc.local which guarantees
that nailgun-agent doesn't change mcollective config at the
same time when cloud-init does this. Otherwise, we can
end up with corrupted mcollective config.

Former location of this file /var/run/nodiscover assumes
it is removed at the beginning of the boot process while
we need it to be in place at least until cloud-init
has finished its magic. For details see
http://www.pathname.com/fhs/pub/fhs-2.3.html#VARRUNRUNTIMEVARIABLEDATA

Change-Id: I5fca4ca39abdc273f6958f2643fae6c5c99fba68
Related-Bug: #1449186