Update erase_node deletion process
When mcollective is launched via Upstart, the current implementation
of the filesystem erase code is not being executed as part of the
SIGTERM handler that is called when we attempt to SIGTERM all
processes. This change pulls the majority of the code we use to
erase a system out of a child thread into the main agent action. This
change leaves the reboot in a child process and SIGTERM handler so
that astute will get the response for the erase_node command and will
mark it complete. An added benefit of this change is that the debug
messages are now returned back to astute and show up in the astute.log
where previously they were being silently discared.

Change-Id: Ib9ee7db0a4d872ea5d1485ec036940f12b0d251e
Closes-Bug: 1461074