Adjust wait to allow erase_node to send response
This change changes the way the erase_node action performs the cleanup
of the disks. We attempt to remount any ext2/3/4 devices as RO and
set the error action to continue. Additionally we are no longer using
sysrq to sync all file systems, reboot all file systems, or sending a
SIGTERM to all processes. Now the disk erase occurs and a sysrq to
reboot the machine after 5 seconds via a forked process. This should
allow for mcollective to send back the response to the fuel node to
acknowledge the erase_node action.

Change-Id: If5ce3d1c2ce99ec422a372bfc30b273583f9fb38
Closes-Bug: 1461074