Adjust wait to allow erase_node to send response
This change adjusts when the SIGTERM is sent to allow the erase_node
action time to respond back to the server. Also being added is a call
to set the error-behavior on any ext2/3/4 filesystem to continue so
that it will continue if a filesystem error is detected. This will
prevent accidently kernel panics and reboots if the filesystem had
the error-behavior set to panic.

Change-Id: If5ce3d1c2ce99ec422a372bfc30b273583f9fb38
Closes-Bug: 1461074