Remove nodes from Cobbler MACs before provisioning
The issue is that when we massively remove nodes from
Cobbler some of them can stay there. It is because
Cobbler is not intended to be scalable. It stores
all systems in plain text files and manipulates these
data quite slowly. So, we need to make sure that
there are no nodes in the Cobbler with the same MAC
addresses, otherwise Cobbler throws MAC address
duplication error.

Change-Id: I822a086c364fc026cd07b2fad4e1810c274d6357
Closes-Bug: #1491725