Quote passwords in astute.yaml right before upload it to nodes
According to the yaml specification and default handling of strings
in yaml files, there is no correct way to force quote strings, so we're
use text patching using regexp right inside text data.

This commit should be compatible with previous MOS versions, without
need to fix this bug in the several places like Nailgun and Hiera.

Change-Id: I7435da01016c4b2266210eb9422ea1fae87de7f8
Closes-Bug: #1494779