Change shotgun command format
Shotgun now provides CLI like:

shotgun snapshot --input-file /path/to/snapshot-config.yaml
shotgun report --input-file /path/to/report-config.yaml

This was introduced to make it possible
to easily gather a short report about the state of the
master node. This change is a part of version.yaml
deprecation plan.

Depends-On: I3bc1f711d0e0fcf06f9414df1533d641a0e46595
Change-Id: Ia1c8034ff3adb5e6f68883f4c263ea07b8dda241
Related-Bug: #1528599