Support rerunning successful puppet tasks
This patch adds the ability to configure astute to retry successful
puppet runs immediately after a successful run. This will be used to
determine the idempotency of a given task. By default it is disabled,
but can be enabled by configuring puppet_succeed_retries in the astute
configuration file.

It will work for both methods: granular and task based deployments

Change-Id: If5af328b71067501cc2cb8d117e78287c1dcf1ad
Related-Blueprint: granular-task-idempotency