Moved parameters of tasks from graph to separate dict
Since the task parameters is same for all nodes, the parameters was
excluded from graph. this parameters is passed to orcestrator in seperate
argument. the graph contains only links between tasks.
This helps to reduce size of graph on more than 3 times.
Astute will construct full graph of tasks based on those 2 arguments

Change-Id: I7e46946c68614789e6c59cc0a0b6ac01fbe6547b
Implements: blueprint reduce-size-of-serialized-tasks