Send cluster id for pre provision task obviously
Orchestrator implements a method called `calculate_cluster_id', which
extracts cluster id from the log file name that is supplied in the `cmd'
field of the parameters hash from pre_provision property of provisioning
info. We should get cluster id in a separate field (`cluster_id')
instead of inferring it from the string representation of the command
and fallback to the old scheme if the `cluster_id' parameter is missing.

Change-Id: I85aebca15c8f3cd7f12760883f2062b7338ed51c
Partial-Bug: #1449512