Erase only boot code in MBR during node delete
Zeroing first 1Mb of a block device will almost always
lead to serios system error, if there is a filesystem
on this block device and this filesystem is mounted.

This code had 2 points: to prevent system from booting
from hard drive after reboot and to prevent anaconda
(or other OS installer) to automatically recognize
disk structure. But since we only use image-based provision
this is not the case anymore.

It's enough to clear first 446 bytes of MBR where the first
stage of a bootloader resides. So we will also be able to
check nodes disks and filesystems after deleting it from
cluster for troubleshooting for example, because partition
scheme will not be touched.

https://wiki.archlinux.org/index.php/Master_Boot_Record

Change-Id: I60f2b250d6d5997b4b471c27ed00d057ae73c30a
Fixes-bug: #1538587