wait until node boot completely before reboot
Reboot hook is used to reboot node with bootstrap image
or provisioned node. It's possible that this hook is called
when boot process of provisioned node haven't completed.
Upstart system doesn't handle such situation and node may
hang.

So let's reboot immediately if we're in a bootstrap. Wait until
system boots completely in case of provisioned node. We check it by
existense of /run/cloud-init/status.json (it's located on tmpfs, so
no stale file from previous boot can be found). If this file hasn't
appeared after 60 seconds - reboot as is.

Partial-Bug: #1573105

Change-Id: Iae113425a0034bc137c88db305d405f79807a5dc