Increase reboot timeout to 15 mins
This change increases the reboot timeout setting from 240 seconds to
900 seconds as some hardware takes much longer to reboot than others.

Change-Id: I4d17e6c1800f108747c50eb7f29638d28ecb555e
Closes-Bug: #1594882
DocImpact: The default timeout for astute to wait for a node to be
rebooted is now 15 minutes. To manually change this, you can update the
astute.conf file and change the reboot_timeout value to the number of
seconds to wait for a reboot. After updating the configuration file, you
will need to restart the service.