Fix encoding of shotgun stdout
Mcollective uses sustemu module to execute shell commands. This module
doesn't care about encoding and as a result we get shotgun stdout as a
string with ASCII-8BIT encoding. Later we get a failure at the attempt
to encode it to JSON. However, shotgun output depends on locale settings
and is usually UTF-8.

Change-Id: I838f1aa55ea2885c1353a352c537dfe1bcd1e7e3
Closes-Bug: #1604311