Add basemodulepath to the Puppet mcagent
This parameter is required for Puppet 4
to be able to find the base modules without
any environment defined.

Puppet 3 is able to work in the legacy mode
without the environment support and does not
case if there is base module path or not.

Closes-Bug: 1655663
Change-Id: I60f2c78ef5fe366314eea186f4671d198e54f1d6
(cherry picked from commit 1c1578b64a5da7f41e74dee35db9a4d3d4f112ed)